# docs/tractive-openapi.yaml
openapi: 3.0.3
info:
  title: Tractive (non officiel)
  version: "2026-05-08"
  description: |
    Spécification OpenAPI **non officielle** de l'API Tractive, restreinte aux
    endpoints réellement consommés par l'application Fire TV `TOT — Tractive On TV`
    (cf. CLAUDE.md). Reconstruite par capture réseau de `my.tractive.com` (web)
    et de l'app iOS officielle (via mitmproxy), croisée avec les implémentations
    communautaires.

    Tous les endpoints sous `/4/` exigent les headers `Authorization`,
    `X-Tractive-User`, `X-Tractive-Client` (sauf `/4/auth/token`). Les valeurs
    des exemples sont neutralisées (`<bearer-token>`, `<user-id>`).

    **Endpoints Wellness (`aps-api.tractive.com`)** — observés en mai 2026 :
    réservés aux comptes Premium, données disponibles après ~7 j d'activité
    (statut `CALCULATING_BASELINE` en attendant). Confirmés joignables avec
    le client web standard, pas besoin du client iOS dédié.
servers:
  - url: https://graph.tractive.com
    description: API principale
  - url: https://aps-api.tractive.com
    description: Wellness / santé
security:
  - bearerAuth: []
    tractiveUser: []
    tractiveClient: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Token opaque obtenu via `POST /4/auth/token`.
    tractiveUser:
      type: apiKey
      in: header
      name: X-Tractive-User
      description: ID utilisateur Tractive (24 hex). Doit correspondre au token.
    tractiveClient:
      type: apiKey
      in: header
      name: X-Tractive-Client
      description: ID client app. Valeur web officielle = `5728aa1fc9077f7c32000186`.
  parameters:
    XTractiveUser:
      name: X-Tractive-User
      in: header
      required: true
      schema: { type: string, example: "<user-id>" }
    XTractiveClient:
      name: X-Tractive-Client
      in: header
      required: true
      schema: { type: string, example: "5728aa1fc9077f7c32000186" }
  schemas:
    AuthTokenRequest:
      type: object
      required: [grant_type, platform_email, platform_token]
      properties:
        grant_type:
          type: string
          enum: [tractive]
        platform_email: { type: string, format: email }
        platform_token: { type: string, description: "Mot de passe utilisateur" }
    AuthToken:
      type: object
      required: [user_id, client_id, expires_at, access_token]
      properties:
        user_id:      { type: string, example: "<user-id>" }
        client_id:    { type: string, example: "5728aa1fc9077f7c32000186" }
        expires_at:   { type: integer, format: int64, description: "Epoch seconds" }
        access_token: { type: string, example: "<bearer-token>" }
    TrackerRef:
      type: object
      required: [_type, _id]
      properties:
        _type: { type: string, enum: [tracker] }
        _id:   { type: string, example: "TRACKER01" }
    TrackerCommandStateRef:
      type: object
      required: [_type, _id]
      properties:
        _type: { type: string, enum: [tracker_command_state] }
        _id:
          type: string
          description: "Format `{trackerIdLower}_{led_control|buzzer_control|live_tracking}`"
          example: "tracker01_live_tracking"
    TrackerCommandState:
      type: object
      properties:
        _id:          { type: string, example: "tracker01_live_tracking" }
        _version:     { type: string }
        _type:        { type: string, enum: [tracker_command_state] }
        active:       { type: boolean }
        started_at:   { type: integer, format: int64, nullable: true }
        timeout:      { type: integer, description: "Durée max en secondes" }
        remaining:    { type: integer, description: "Secondes restantes" }
        pending:      { type: boolean }
        reconnecting: { type: boolean, description: "Présent uniquement pour live_tracking" }
    DevicePosReport:
      type: object
      properties:
        _id:         { type: string }
        _type:       { type: string, enum: [device_pos_report] }
        time:        { type: integer, format: int64 }
        latlong:
          type: array
          minItems: 2
          maxItems: 2
          items: { type: number }
        pos_uncertainty: { type: number }
        sensor_used:     { type: string, enum: [GPS, WIFI, CELLULAR, KNOWN_WIFI, BLUETOOTH] }
    DeviceHwReport:
      type: object
      properties:
        _id:                  { type: string }
        _type:                { type: string, enum: [device_hw_report] }
        battery_level:        { type: integer }
        tracker_state:        { type: string }
        power_saving_zone_id: { type: string, nullable: true }
        charging_state:       { type: string }
    BulkRequest:
      type: array
      items:
        oneOf:
          - $ref: "#/components/schemas/TrackerRef"
          - $ref: "#/components/schemas/TrackerCommandStateRef"
      example:
        - { _type: tracker_command_state, _id: "tracker01_live_tracking" }
        - { _type: tracker, _id: "TRACKER01" }
    BulkResponse:
      type: array
      description: |
        Tableau hétérogène d'objets résolus. Types observés : `tracker`,
        `tracker_command_state`, `device_pos_report`, `device_hw_report`, `pet`.
      items:
        oneOf:
          - $ref: "#/components/schemas/TrackerCommandState"
          - $ref: "#/components/schemas/DevicePosReport"
          - $ref: "#/components/schemas/DeviceHwReport"
          - $ref: "#/components/schemas/Pet"
...
    Geofence:
      type: object
      properties:
        _id:      { type: string }
        _type:    { type: string, enum: [geofence] }
        _version: { type: string }
        name:     { type: string }
        active:   { type: boolean }
        shape:
          type: object
          properties:
            type:        { type: string, enum: [polygon, circle] }
            coordinates:
              type: array
              description: Polygone fermé [[lng, lat], …] ou centre+rayon selon `type`.
              items:
                type: array
                items: { type: number }
    Pet:
      type: object
      description: |
        Animal suivi. Le `_type` réel observé est **`pet`** (pas
        `trackable_object` comme suggéré par le nom de la route).
      required: [_id, _type]
      properties:
        _id:        { type: string, example: "000000000000000000000000" }
        _type:      { type: string, enum: [pet] }
        _version:   { type: string }
        device_id:  { type: string, description: "ID du tracker associé", example: "TRACKER01" }
        details:
          type: object
          properties:
            name:     { type: string, example: "Felix" }
            pet_type: { type: string, enum: [CAT, DOG] }
            gender:   { type: string, enum: [F, M, U] }
            birthday: { type: integer, format: int64 }
            weight:   { type: integer, nullable: true, description: "Poids en grammes" }
            profile_picture_id: { type: string, nullable: true }
...
  responses:
    Unauthorized:
      description: Token absent ou invalide
      content:
        application/json:
          schema:
            type: object
            properties:
              error:   { type: string }
              message: { type: string }
paths:
...
  /4/auth/token:
    post:
      tags: [Auth]
      summary: Connexion (obtenir un access_token)
      security: []
      parameters:
        - $ref: "#/components/parameters/XTractiveClient"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AuthTokenRequest" }
      responses:
        "200":
          description: Token émis
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthToken" }
        "401":
          $ref: "#/components/responses/Unauthorized"

  # --- Trackers ------------------------------------------------------------
...
  /4/tracker/{trackerId}/geofences:
    get:
      tags: [Tracker]
      summary: Liste des zones de sécurité
      parameters:
        - { name: trackerId, in: path, required: true, schema: { type: string } }
        - $ref: "#/components/parameters/XTractiveUser"
        - $ref: "#/components/parameters/XTractiveClient"
      responses:
        "200":
          description: Liste de geofences
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Geofence" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /4/bulk:
    post:
      tags: [Bulk]
      summary: Lecture en lot d'objets graph
      description: |
        Endpoint principal du polling 60 s (mode standard) et 20 s (mode live).
        Permet de récupérer en un seul appel batterie + position + statut
        commande live. Types attendus en requête : `tracker`,
        `tracker_command_state`. Types retournés : `tracker`,
        `tracker_command_state`, `device_pos_report`, `device_hw_report`, `pet`.
      parameters:
        - { name: partial, in: query, required: false, schema: { type: integer, enum: [0, 1] }, description: "1 = autorise les réponses partielles" }
        - $ref: "#/components/parameters/XTractiveUser"
        - $ref: "#/components/parameters/XTractiveClient"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkRequest" }
      responses:
        "200":
          description: Tableau d'objets résolus
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BulkResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"

  # --- Wellness ------------------------------------------------------------
...